imageviewer improvements:  rotation by TAB/SHIFT+TAB  SHIFT with other keys reduces zoom/offset step  RGB support instead of only RGBA to save bandwidth  some other improvements for smoother UX

+++ WinPort/WinPortDecl.h
@@ -80,7 +80,8 @@
 
 	// graphics API
 	WINPORT_DECL_DEF(GetConsoleImageCaps, BOOL, (HANDLE hConsoleOutput, size_t sizeof_wgi, WinportGraphicsInfo *wgi))
-	WINPORT_DECL_DEF(SetConsoleImage, BOOL, (HANDLE hConsoleOutput, const char *id, DWORD flags, COORD pos, DWORD width, DWORD height, const void *buffer))
+	// flags: format identity, one of WP_IMG_ that is supported according to GetConsoleImageCaps
+	WINPORT_DECL_DEF(SetConsoleImage, BOOL, (HANDLE hConsoleOutput, const char *id, DWORD64 flags, COORD pos, DWORD width, DWORD height, const void *buffer))
 	WINPORT_DECL_DEF(DeleteConsoleImage, BOOL, (HANDLE hConsoleOutput, const char *id))
 
 #ifdef WINPORT_REGISTRY
